requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
flatbuffers==24.3.25
gast==0.4.0
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.68.0
h5py==3.12.1
idna==3.10
importlib_metadata==8.5.0
keras==2.13.1
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.2
numpy==1.24.3
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==24.2
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydot==1.4.2
pyparsing==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow-macos==2.13.0
tensorflow-metal==1.0.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.5.0
typing_extensions==4.5.0
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
zipp==3.21.0
opencv-python==4.10.0.84